Outsource services add error handling for copy broken files.
