+++ src/services/lesson/services/index.js
@@ -1,0 +1,7 @@
+const LessonCopyService = require('./LessonCopyService');
+const LessonFilesService = require('./LessonFilesService');
+
+module.exports = {
+	LessonCopyService,
+	LessonFilesService,
+};
